Disable calling to get_measures_list in the get_measures rest call
