+++ gnocchi/tests/test_amqp1d.py
@@ -11,7 +11,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 import datetime
-import itertools
 import json
 import uuid
 
@@ -19,26 +18,13 @@ import mock
 import numpy
 
 from gnocchi import amqp1d
-from gnocchi import storage
 from gnocchi.tests import base as tests_base
+from gnocchi.tests.test_utils import get_measures_list
 from gnocchi import utils
 
 
 def datetime64(*args):
     return numpy.datetime64(datetime.datetime(*args))
-
-
-def get_measures_list(measures_agg):
-    return {
-        aggmethod: list(itertools.chain(
-            *[[(timestamp, measures_agg[agg].aggregation.granularity, value)
-               for timestamp, value in measures_agg[agg]]
-              for agg in sorted(aggs,
-                                key=storage.ATTRGETTER_GRANULARITY,
-                                reverse=True)]))
-        for aggmethod, aggs in itertools.groupby(measures_agg.keys(),
-                                                 storage.ATTRGETTER_METHOD)
-    }
 
 
 class TestAmqp1d(tests_base.TestCase):
